requirements.txt
streamlit==1.44.0
openai==1.69.0
google-generativeai==0.8.4
langchain==0.3.21
langchain-google-genai==2.0.10
pypdf==5.4.0
python-dotenv==1.1.0
markdown==3.7
langchain-openai==0.3.16